update all gallery routes to use same spacing component
