+++ pages/[username]/index.tsx
@@ -1,11 +1,15 @@
 import { GetServerSideProps } from 'next';
 import { route } from 'nextjs-routes';
+import { PropsWithChildren } from 'react';
 import { useLazyLoadQuery } from 'react-relay';
 import { fetchQuery, graphql } from 'relay-runtime';
+import styled from 'styled-components';
 
+import breakpoints, { pageGutter } from '~/components/core/breakpoints';
 import useVerifyEmailOnPage from '~/components/Email/useVerifyEmailOnPage';
 import GalleryViewEmitter from '~/components/internal/GalleryViewEmitter';
 import { GalleryNavbar } from '~/contexts/globalLayout/GlobalNavbar/GalleryNavbar/GalleryNavbar';
+import { useGlobalNavbarHeight } from '~/contexts/globalLayout/GlobalNavbar/useGlobalNavbarHeight';
 import { UsernameQuery } from '~/generated/UsernameQuery.graphql';
 import { MetaTagProps } from '~/pages/_app';
 import GalleryRoute from '~/scenes/_Router/GalleryRoute';
@@ -21,6 +25,38 @@ const UsernameQueryNode = graphql`
     ...useOpenSettingsModalFragment
     ...GalleryViewEmitterWithSuspenseFragment
     ...useVerifyEmailOnPageQueryFragment
+  }
+`;
+
+// This component exists here since all child routes of this page (/{username}/*)
+// will be using this wrapper.
+// In the future (Next 13 App Directory), we can just use a layout file
+export function GalleryPageSpacing({ children }: PropsWithChildren) {
+  const navbarHeight = useGlobalNavbarHeight();
+
+  return (
+    <GalleryPageSpacingContainer navbarHeight={navbarHeight}>
+      <GalleryPageSpacingInner>{children}</GalleryPageSpacingInner>
+    </GalleryPageSpacingContainer>
+  );
+}
+
+const GalleryPageSpacingInner = styled.div`
+  max-width: 1200px;
+  width: 100%;
+`;
+
+const GalleryPageSpacingContainer = styled.div<{ navbarHeight: number }>`
+  display: flex;
+  justify-content: center;
+  min-height: 100vh;
+
+  margin: 0 ${pageGutter.mobile}px 24px;
+  padding-top: ${({ navbarHeight }) => navbarHeight + 10}px;
+
+  @media only screen and ${breakpoints.tablet} {
+    margin: 0 ${pageGutter.tablet}px;
+    padding-top: ${({ navbarHeight }) => navbarHeight + 24}px;
   }
 `;
