+++ src/scenes/UserGalleryPage/FocusedGalleryPage.tsx
@@ -8,6 +8,7 @@ import breakpoints, { pageGutter } from '~/components/core/breakpoints';
 import { useTrack } from '~/contexts/analytics/AnalyticsContext';
 import { useGlobalNavbarHeight } from '~/contexts/globalLayout/GlobalNavbar/useGlobalNavbarHeight';
 import { FocusedGalleryPageFragment$key } from '~/generated/FocusedGalleryPageFragment.graphql';
+import { GalleryPageSpacing } from '~/pages/[username]';
 import { UserGalleryLayout } from '~/scenes/UserGalleryPage/UserGalleryLayout';
 
 type FocusedGalleryPageProps = {
@@ -59,23 +60,9 @@ export function FocusedGalleryPage({ queryRef }: FocusedGalleryPageProps) {
       <Head>
         <title>{headTitle}</title>
       </Head>
-      <StyledUserGalleryPage navbarHeight={navbarHeight}>
+      <GalleryPageSpacing>
         <UserGalleryLayout galleryRef={query.galleryById} queryRef={query} />
-      </StyledUserGalleryPage>
+      </GalleryPageSpacing>
     </>
   );
 }
-
-export const StyledUserGalleryPage = styled.div<{ navbarHeight: number }>`
-  display: flex;
-  justify-content: center;
-  min-height: 100vh;
-
-  margin: 0 ${pageGutter.mobile}px 24px;
-  padding-top: ${({ navbarHeight }) => navbarHeight + 10}px;
-
-  @media only screen and ${breakpoints.tablet} {
-    margin: 0 ${pageGutter.tablet}px;
-    padding-top: ${({ navbarHeight }) => navbarHeight + 24}px;
-  }
-`;
