+++ src/scenes/UserGalleryPage/UserGalleryLayout.tsx
@@ -53,7 +53,7 @@ export const UserGalleryLayout = ({ galleryRef, queryRef }: Props) => {
   );
 
   return (
-    <StyledGalleryLayout>
+    <VStack>
       <GalleryNameDescriptionHeader
         noLink={pathname === '/[username]/galleries/[galleryId]'}
         galleryRef={gallery}
@@ -64,11 +64,6 @@ export const UserGalleryLayout = ({ galleryRef, queryRef }: Props) => {
       <VStack gap={32} align="center" grow>
         {collectionsView}
       </VStack>
-    </StyledGalleryLayout>
+    </VStack>
   );
 };
-
-export const StyledGalleryLayout = styled(VStack)`
-  width: 100%;
-  padding-bottom: 32px;
-`;
