+++ src/scenes/UserGalleryPage/UserGalleryPage.tsx
@@ -8,6 +8,7 @@ import breakpoints, { pageGutter } from '~/components/core/breakpoints';
 import { useTrack } from '~/contexts/analytics/AnalyticsContext';
 import { useGlobalNavbarHeight } from '~/contexts/globalLayout/GlobalNavbar/useGlobalNavbarHeight';
 import { UserGalleryPageFragment$key } from '~/generated/UserGalleryPageFragment.graphql';
+import { GalleryPageSpacing } from '~/pages/[username]';
 
 import UserGallery from './UserGallery';
 
@@ -29,7 +30,6 @@ function UserGalleryPage({ queryRef, username }: UserGalleryPageProps) {
   const headTitle = `${username} | Gallery`;
 
   const track = useTrack();
-  const navbarHeight = useGlobalNavbarHeight();
 
   useEffect(() => {
     track('Page View: User Gallery', { username }, true);
@@ -40,24 +40,11 @@ function UserGalleryPage({ queryRef, username }: UserGalleryPageProps) {
       <Head>
         <title>{headTitle}</title>
       </Head>
-      <StyledUserGalleryPage navbarHeight={navbarHeight}>
+      <GalleryPageSpacing>
         <UserGallery queryRef={query} />
-      </StyledUserGalleryPage>
+      </GalleryPageSpacing>
     </>
   );
 }
-
-export const StyledUserGalleryPage = styled.div<{ navbarHeight: number }>`
-  display: flex;
-  justify-content: center;
-  padding-top: ${({ navbarHeight }) => navbarHeight}px;
-  min-height: 100vh;
-
-  margin: 0 ${pageGutter.mobile}px 24px;
-
-  @media only screen and ${breakpoints.tablet} {
-    margin: 0 ${pageGutter.tablet}px;
-  }
-`;
 
 export default UserGalleryPage;
